get-started/requirements-pyg.txt
# Install pre-built wheels for PyTorch Geometric that are compatible with 
# poptorch
--find-links https://data.pyg.org/whl/torch-1.10.0+cpu.html

torch-cluster==1.6.0
torch-scatter==2.0.9
torch-sparse==0.6.13
torch-spline-conv==1.2.1
# TODO replace with pyg-nightly build once this PR is merged:
#https://github.com/pyg-team/pytorch_geometric/pull/6100
torch-geometric@git+https://github.com/graphcore/pytorch_geometric-fork@validate-property

# Visualisation packages
pandas==1.5.0
matplotlib==3.6.0
periodictable==1.6.1
py3Dmol==1.8.1
seaborn==0.12.1